cutting dependencies on vscode
