+++ extension/src/ALObject/ALParser.ts
@@ -12,7 +12,6 @@ import { ALPagePart } from "./ALPagePart";
 import { ALProcedure } from "./ALProcedure";
 import { ALTableField } from "./ALTableField";
 import { ALXmlComment } from "./ALXmlComment";
-import * as DocumentFunctions from "../DocumentFunctions";
 import * as fs from "fs";
 import {
   ALControlType,
@@ -569,7 +568,7 @@ export function getALObjectFromText(
       0
     );
     if (objectAsText) {
-      alObj.eol = DocumentFunctions.getEOL(objectAsText);
+      alObj.eol = ALObject.getEOL(objectAsText);
     }
   }
   if (alObjects) {
