+++ extension/src/DocumentFunctions.ts
@@ -1,5 +1,9 @@
 import * as vscode from "vscode";
 import * as fs from "fs";
+import * as ALParser from "./ALObject/ALParser";
+import { XliffIdToken } from "./ALObject/XliffIdToken";
+import { AppManifest, Settings } from "./Settings";
+import { getAlObjectsFromCurrentWorkspace } from "./WorkspaceFunctions";
 
 export async function openTextFileWithSelection(
   documentFilePath: string,
@@ -63,10 +67,47 @@ export function eolToLineEnding(eol: vscode.EndOfLine): string {
   return "\n";
 }
 
-export function getEOL(source: string): vscode.EndOfLine {
-  const temp = source.indexOf("\n");
-  if (source[temp - 1] === "\r") {
-    return vscode.EndOfLine.CRLF;
+export async function openAlFileFromXliffTokens(
+  settings: Settings,
+  appManifest: AppManifest,
+  tokens: XliffIdToken[]
+): Promise<void> {
+  const alObjects = await getAlObjectsFromCurrentWorkspace(
+    settings,
+    appManifest,
+    false
+  );
+  const obj = alObjects.filter(
+    (x) =>
+      x.objectType.toLowerCase() === tokens[0].type.toLowerCase() &&
+      x.objectName.toLowerCase() === tokens[0].name.toLowerCase()
+  )[0];
+  if (!obj) {
+    throw new Error(
+      `Could not find any object matching '${XliffIdToken.getXliffIdWithNames(
+        tokens
+      )}'`
+    );
   }
-  return vscode.EndOfLine.LF;
+  // found our object, load complete object from file
+  obj.endLineIndex = ALParser.parseCode(obj, obj.startLineIndex + 1, 0);
+
+  const xliffToSearchFor = XliffIdToken.getXliffId(tokens).toLowerCase();
+  const mlObjects = obj.getAllMultiLanguageObjects({
+    onlyForTranslation: true,
+  });
+  const mlObject = mlObjects.filter(
+    (x) => x.xliffId().toLowerCase() === xliffToSearchFor
+  );
+  if (mlObject.length !== 1) {
+    throw new Error(
+      `No code line found in file '${
+        obj.objectFileName
+      }' matching '${XliffIdToken.getXliffIdWithNames(tokens)}'`
+    );
+  }
+  openTextFileWithSelectionOnLineNo(
+    obj.objectFileName,
+    mlObject[0].startLineIndex
+  );
 }
